git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@62 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/MathLib.cpp
@@ -36,7 +36,7 @@ float fastSQRT(float x)
     if( x == 0 )
         return 0;
     return 1.0f / x; */
-	return sqrt(x);
+	return (float)sqrt(x);
 }
